Add game-api functions
This commit adds the functions to accomplish the three primary game
interactions with the API: create game, patch game, and get games.
Create adds a new game record to the API, get games will eventually
return a list of all of the user's previous and current games, and patch
game will later be used to update the games' board states in the API.
This commit marks where the functions are successfully connecting with
the API and not where they are finalized and with their full, intended
functionality. The get functions does not print anything for the user at
this time and the patch functions will not be complete until game logic
has been implemented.

Additionally, the HTML has been updated accordingly with three
placeholder buttons that are being used to invoke and test the create,
get, and patch functions.