Add game reset and further game-get features
This commit adds a Game Reset button to the HTML. Clicking it resets the
text in each game cell to the default '-' and creates a new game.
Currently, the Get Game button will return all of the player's games
when clicked. This is not formatted at this time, so the list is
unintelligible right now.

With this commit, the majority of the MVP's functionality is present and
working as intended. Further necessary changes include hiding irrelevant
forms (e.g. the Sign In form when the user is already signed in). Other
future updates may include reformatting the Get Games button's output to
be more legible and/or improving the app's visual design.