Reorganize board and add Get function
This commit adds full functionality to the "Get Games" button, which
now returns the number of total games played by the user. Additionally,
the app's layout has been updated to center align certain titles and
messages, as well as move buttons and messages below the board. Lastly,
the center board cell now contains instructions for the user to start
playing, namely "Sign in" and "Click 'New Game.'" These are removed when
'New Game' is clicked and a game starts.