Add /game/ repository and api, events, ui files
This commit adds a new /game/ repository to /assets/scripts/. This
includes an api.js, events.js, and ui.js files, which are analgous to
the /auth/'s files of the same name. Code related to game functions (e.g
creating, patching, or getting games from the API) have been moved to
these files to better organize the app and differentiate between
authorization and game-related functions. With this update, there was
miscellaneous changes throughout the files to adjust for the file path
updates.