Fix bug on-click bug and add new messages
This commit fixes a bug where the on-click event for the board logged a
console error if the board was clicked before creating a new game. Now,
the same action prints a new message to the app. Additionally, there is
now a similar message for if a user clicks on the board after the game
is over to alert them.