Fix sign-out bug
Before now, if a user filled in both or either of the Change Password
fields and then signed out without submitting the Change Password form,
the fields would not clear, letting a new user sign in and see the
filled-in fields. This has been fixed by clearing those fields on
submission of the Sign Out button.